# docs/swagger.yaml
definitions:
...
  job_sms.CreateRequest:
    properties:
      message:
        type: string
      recipients:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  job_sms.ItemResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      jobId:
        type: integer
      message:
        type: string
      recipients:
        items:
          type: string
        type: array
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  job_task.SendNotificationRequest:
    properties:
      email:
        type: string
    type: object
...
  quote.SendEmailRequest:
    properties:
      email:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  response.PaginatedResponse:
    properties:
      items: {}
      page:
        type: integer
      pageSize:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
...
host: localhost:8090
info:
  contact:
    email: support@jvair.com
    name: API Support
    url: http://www.jvair.com/support
  description: API para el sistema JVAIR
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: JVAIR API
  version: "1.0"
paths:
...
  /api/v1/jobs/{jobId}/sms:
    get:
      consumes:
      - application/json
      description: Obtiene una lista paginada de SMS enviados asociados a un trabajo
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: integer
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 15
        description: Tamaño de página
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar SMS enviados de un trabajo
      tags:
      - JobSMS
    post:
      consumes:
      - application/json
      description: Crea un registro de SMS enviado asociado a un trabajo
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: integer
      - description: Datos del SMS enviado
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/job_sms.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/job_sms.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Registrar SMS enviado de un trabajo
      tags:
      - JobSMS
  /api/v1/jobs/{jobId}/sms/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un registro de SMS enviado por ID
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: integer
      - description: ID del SMS
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Eliminar SMS enviado de un trabajo
      tags:
      - JobSMS
...
  /quotes/{id}/email:
    post:
      consumes:
      - application/json
      description: Send quote email to specified recipients
      parameters:
      - description: Quote ID
        in: path
        name: id
        required: true
        type: integer
      - description: Email request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/quote.SendEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send quote email
      tags:
      - quotes
  /tasks/{id}/notification:
    post:
      consumes:
      - application/json
      description: Send task notification email to specified recipients
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Email request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/job_task.SendNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send task notification email
      tags:
      - tasks
...
securityDefinitions:
  BearerAuth:
    description: Ingrese el token con el prefijo 'Bearer '
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...